Update twosum_concept.c, redid weird choice
I remembered why I did that now.

numsSize doesn't seem to actually give the number of integers in the array? Trying to use it in this way ends up with memory issues. I have no idea what numsSize actually does, because the TwoSum description does not tell me.

So in order to calculate the number of integers in the array, I calculate the memory size of the whole array and divide it by the memory size of the first integer(?) of the array.

But! This does not seem to work still.